requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==1.10.8
sqlalchemy==1.4.49
python-dotenv==1.0.0
flask==2.3.3
gunicorn==21.2.0
